fix: separate router for changing email
